circuits: zk-gadgets: bitlength: Add bitlength gadgets
