+++ circuits/src/zk_gadgets/v2/mod.rs
@@ -1,5 +1,6 @@
 //! V2 gadgets for zero knowledge circuits
 
+pub mod bitlength;
 pub mod csprng;
 pub mod shares;
 pub mod state_elements;
